♻️ refactor:"redis.util 수정"
updateRoom 하면 return 값으로 업데이트한 방 정보 반환하게 수정
