+++ game.server/src/managers/socket.manger.ts
@@ -1,4 +1,3 @@
-import { Socket } from 'net';
 import { GameSocket } from '../type/game.socket';
 
 // 전역적으로 연결된 소켓들을 관리하는 Map (userId를 키로 사용)
